Improve Task adapter and its utils
